requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
vl-convert-python==1.9.0.post1
